LPS-100078 Refactor tests to split into separate archetype projects
