+++ modules/sdk/project-templates/project-templates-form-field/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
