+++ modules/sdk/project-templates/project-templates-fragment/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplatesArgsExt
