+++ modules/sdk/project-templates/project-templates-fragment/src/test/java/com/liferay/project/templates/fragment/ProjectTemplatesFragmentTest.java
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.project.templates.fragment;
+
+import aQute.bnd.header.Attrs;
+import aQute.bnd.osgi.Domain;
+
+import com.liferay.maven.executor.MavenExecutor;
+import com.liferay.project.templates.BaseProjectTemplatesTestCase;
+import com.liferay.project.templates.extensions.util.Validator;
+import com.liferay.project.templates.util.FileTestUtil;
+
+import java.io.File;
+
+import java.net.URI;
+
+import java.util.Map;
+import java.util.Properties;
+
+import org.junit.Assert;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+/**
+ * @author Gregory Amerson
+ */
+public class ProjectTemplatesFragmentTest
+	implements BaseProjectTemplatesTestCase {
+
+	@ClassRule
+	public static final MavenExecutor mavenExecutor = new MavenExecutor();
+
+	@BeforeClass
+	public static void setUpClass() throws Exception {
+		String gradleDistribution = System.getProperty("gradle.distribution");
+
+		if (Validator.isNull(gradleDistribution)) {
+			Properties properties = FileTestUtil.readProperties(
+				"gradle-wrapper/gradle/wrapper/gradle-wrapper.properties");
+
+			gradleDistribution = properties.getProperty("distributionUrl");
+		}
+
+		Assert.assertTrue(gradleDistribution.contains(GRADLE_WRAPPER_VERSION));
+
+		_gradleDistribution = URI.create(gradleDistribution);
+	}
+
+	@Test
+	public void testBuildTemplateFragment() throws Exception {
+		File gradleProjectDir = _buildTemplateWithGradle(
+			"fragment", "loginhook", "--host-bundle-symbolic-name",
+			"com.liferay.login.web", "--host-bundle-version", "1.0.0");
+
+		testContains(
+			gradleProjectDir, "bnd.bnd", "Bundle-SymbolicName: loginhook",
+			"Fragment-Host: com.liferay.login.web;bundle-version=\"1.0.0\"");
+		testContains(
+			gradleProjectDir, "build.gradle",
+			"apply plugin: \"com.liferay.plugin\"");
+
+		File mavenProjectDir = buildTemplateWithMaven(
+			temporaryFolder, "fragment", "loginhook", "com.test", mavenExecutor,
+			"-DhostBundleSymbolicName=com.liferay.login.web",
+			"-DhostBundleVersion=1.0.0", "-Dpackage=loginhook");
+
+		buildProjects(
+			_gradleDistribution, mavenExecutor, gradleProjectDir,
+			mavenProjectDir);
+
+		if (isBuildProjects()) {
+			File jarFile = testExists(
+				gradleProjectDir, "build/libs/loginhook-1.0.0.jar");
+
+			Domain domain = Domain.domain(jarFile);
+
+			Map.Entry<String, Attrs> fragmentHost = domain.getFragmentHost();
+
+			Assert.assertNotNull(fragmentHost);
+
+			Assert.assertEquals(
+				fragmentHost.toString(), "com.liferay.login.web",
+				fragmentHost.getKey());
+		}
+	}
+
+	@Rule
+	public TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+	private File _buildTemplateWithGradle(
+			String template, String name, String... args)
+		throws Exception {
+
+		return buildTemplateWithGradle(temporaryFolder, template, name, args);
+	}
+
+	private static URI _gradleDistribution;
+
+}
