+++ modules/sdk/project-templates/project-templates-modules-ext/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplatesArgsExt
