+++ modules/sdk/project-templates/project-templates-npm-angular-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
