+++ modules/sdk/project-templates/project-templates-npm-react-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
