+++ modules/sdk/project-templates/project-templates-npm-vuejs-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
