+++ modules/sdk/project-templates/project-templates-rest/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
