+++ modules/sdk/project-templates/project-templates-service-builder/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
