+++ modules/sdk/project-templates/project-templates-service/build.gradle
@@ -2,4 +2,14 @@ dependencies {
 	compileOnly group: "com.beust", name: "jcommander", version: "1.72"
 	compileOnly group: "com.liferay", name: "com.liferay.project.templates.extensions", version: "1.0.0"
 	compileOnly group: "org.apache.maven.archetype", name: "archetype-common", version: "2.4"
+
+	testCompile gradleTestKit()
+	testCompile group: "biz.aQute.bnd", name: "biz.aQute.bnd", version: "3.5.0"
+	testCompile group: "com.googlecode.java-diff-utils", name: "diffutils", version: "1.3.0"
+	testCompile group: "com.liferay", name: "com.liferay.maven.executor", version: "1.0.0"
+	testCompile group: "junit", name: "junit", version: "4.12"
+	testCompile group: "net.diibadaaba.zipdiff", name: "zipdiff", version: "1.0"
+	testCompile project(path: ":sdk:project-templates:project-templates", configuration: "testAPI")
+
+	testRuntime project(":sdk:project-templates:project-templates").files("classes")
 }
