+++ modules/sdk/project-templates/project-templates-service/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
