+++ modules/sdk/project-templates/project-templates-spring-mvc-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplatesArgsExt
