+++ modules/sdk/project-templates/project-templates-spring-mvc-portlet/src/test/java/com/liferay/project/templates/spring/mvc/portlet/ProjectTemplatesSpringPortletMVCTest.java
@@ -12,24 +12,18 @@
  * details.
  */
 
-package com.liferay.project.templates;
+package com.liferay.project.templates.spring.mvc.portlet;
 
 import com.liferay.maven.executor.MavenExecutor;
-import com.liferay.project.templates.internal.util.ProjectTemplatesUtil;
-import com.liferay.project.templates.internal.util.Validator;
+import com.liferay.project.templates.BaseProjectTemplatesTestCase;
+import com.liferay.project.templates.extensions.util.Validator;
 import com.liferay.project.templates.util.FileTestUtil;
 
 import java.io.File;
 
 import java.net.URI;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Properties;
 
 import org.junit.Assert;
@@ -50,10 +44,6 @@ public class ProjectTemplatesSpringPortletMVCTest
 
 	@ClassRule
 	public static final MavenExecutor mavenExecutor = new MavenExecutor();
-
-	@ClassRule
-	public static final TemporaryFolder testCaseTemporaryFolder =
-		new TemporaryFolder();
 
 	@Parameterized.Parameters(
 		name = "Testcase-{index}: testing {0}, {1}, {2}, {3}"
@@ -183,80 +173,13 @@ public class ProjectTemplatesSpringPortletMVCTest
 			"maven", _framework, _frameworkDependencies, _viewType,
 			_liferayVersion);
 
-		_buildProjects(gradleProjectDir, mavenProjectDir);
+		buildProjects(
+			_gradleDistribution, mavenExecutor, gradleProjectDir,
+			mavenProjectDir);
 	}
 
 	@Rule
 	public final TemporaryFolder temporaryFolder = new TemporaryFolder();
-
-	private void _buildProjects(File gradleProjectDir, File mavenProjectDir)
-		throws Exception {
-
-		File gradleOutputDir = new File(gradleProjectDir, "build/libs");
-		File mavenOutputDir = new File(mavenProjectDir, "target");
-
-		_buildProjects(
-			gradleProjectDir, mavenProjectDir, gradleOutputDir, mavenOutputDir,
-			GRADLE_TASK_PATH_BUILD);
-	}
-
-	private void _buildProjects(
-			File gradleProjectDir, File mavenProjectDir, File gradleOutputDir,
-			File mavenOutputDir, String... gradleTaskPath)
-		throws Exception {
-
-		if (isBuildProjects()) {
-			executeGradle(
-				gradleProjectDir, _gradleDistribution, gradleTaskPath);
-
-			Path gradleOutputPath = FileTestUtil.getFile(
-				gradleOutputDir.toPath(), OUTPUT_FILENAME_GLOB_REGEX, 1);
-
-			Assert.assertNotNull(gradleOutputPath);
-
-			Assert.assertTrue(Files.exists(gradleOutputPath));
-
-			File gradleOutputFile = gradleOutputPath.toFile();
-
-			String gradleOutputFileName = gradleOutputFile.getName();
-
-			_executeMaven(mavenProjectDir, MAVEN_GOAL_PACKAGE);
-
-			Path mavenOutputPath = FileTestUtil.getFile(
-				mavenOutputDir.toPath(), OUTPUT_FILENAME_GLOB_REGEX, 1);
-
-			Assert.assertNotNull(mavenOutputPath);
-
-			Assert.assertTrue(Files.exists(mavenOutputPath));
-
-			File mavenOutputFile = mavenOutputPath.toFile();
-
-			String mavenOutputFileName = mavenOutputFile.getName();
-
-			try {
-				if (gradleOutputFileName.endsWith(".jar")) {
-					testBundlesDiff(gradleOutputFile, mavenOutputFile);
-				}
-				else if (gradleOutputFileName.endsWith(".war")) {
-					testWarsDiff(gradleOutputFile, mavenOutputFile);
-				}
-			}
-			catch (Throwable t) {
-				if (TEST_DEBUG_BUNDLE_DIFFS) {
-					Path dirPath = Paths.get("build");
-
-					Files.copy(
-						gradleOutputFile.toPath(),
-						dirPath.resolve(gradleOutputFileName));
-					Files.copy(
-						mavenOutputFile.toPath(),
-						dirPath.resolve(mavenOutputFileName));
-				}
-
-				throw t;
-			}
-		}
-	}
 
 	private File _buildSpringMVCTemplate(
 			String buildType, String framework, String frameworkDependencies,
@@ -269,9 +192,10 @@ public class ProjectTemplatesSpringPortletMVCTest
 		if (buildType.equals("maven")) {
 			String groupId = "com.test";
 
-			return _buildTemplateWithMaven(
-				template, name, groupId, "-Dpackage=com.test",
-				"-DclassName=Sample", "-Dframework=" + framework,
+			return buildTemplateWithMaven(
+				temporaryFolder, template, name, groupId, mavenExecutor,
+				"-Dpackage=com.test", "-DclassName=Sample",
+				"-Dframework=" + framework,
 				"-DframeworkDependencies=" + frameworkDependencies,
 				"-DviewType=" + viewType, "-DliferayVersion=" + liferayVersion);
 		}
@@ -290,133 +214,6 @@ public class ProjectTemplatesSpringPortletMVCTest
 		File destinationDir = temporaryFolder.newFolder("gradle");
 
 		return buildTemplateWithGradle(destinationDir, template, name, args);
-	}
-
-	private File _buildTemplateWithMaven(
-			File parentDir, File destinationDir, String template, String name,
-			String groupId, String... args)
-		throws Exception {
-
-		List<String> completeArgs = new ArrayList<>();
-
-		completeArgs.add("archetype:generate");
-		completeArgs.add("--batch-mode");
-
-		String archetypeArtifactId =
-			"com.liferay.project.templates." + template.replace('-', '.');
-
-		if (archetypeArtifactId.equals(
-				"com.liferay.project.templates.portlet")) {
-
-			archetypeArtifactId = "com.liferay.project.templates.mvc.portlet";
-		}
-
-		completeArgs.add("-DarchetypeArtifactId=" + archetypeArtifactId);
-
-		String projectTemplateVersion =
-			ProjectTemplatesUtil.getArchetypeVersion(archetypeArtifactId);
-
-		Assert.assertTrue(
-			"Unable to get project template version",
-			Validator.isNotNull(projectTemplateVersion));
-
-		completeArgs.add("-DarchetypeGroupId=com.liferay");
-		completeArgs.add("-DarchetypeVersion=" + projectTemplateVersion);
-		completeArgs.add("-Dauthor=" + System.getProperty("user.name"));
-		completeArgs.add("-DgroupId=" + groupId);
-		completeArgs.add("-DartifactId=" + name);
-		completeArgs.add("-Dversion=1.0.0");
-
-		boolean liferayVersionSet = false;
-		boolean projectTypeSet = false;
-
-		for (String arg : args) {
-			completeArgs.add(arg);
-
-			if (arg.startsWith("-DliferayVersion=")) {
-				liferayVersionSet = true;
-			}
-			else if (arg.startsWith("-DprojectType=")) {
-				projectTypeSet = true;
-			}
-		}
-
-		if (!liferayVersionSet) {
-			completeArgs.add("-DliferayVersion=7.2");
-		}
-
-		if (!projectTypeSet) {
-			completeArgs.add("-DprojectType=standalone");
-		}
-
-		_executeMaven(destinationDir, completeArgs.toArray(new String[0]));
-
-		File projectDir = new File(destinationDir, name);
-
-		testExists(projectDir, "pom.xml");
-		testNotExists(projectDir, "gradlew");
-		testNotExists(projectDir, "gradlew.bat");
-		testNotExists(projectDir, "gradle/wrapper/gradle-wrapper.jar");
-		testNotExists(projectDir, "gradle/wrapper/gradle-wrapper.properties");
-
-		testArchetyper(
-			parentDir, destinationDir, projectDir, name, groupId, template,
-			completeArgs);
-
-		return projectDir;
-	}
-
-	private File _buildTemplateWithMaven(
-			String template, String name, String groupId, String... args)
-		throws Exception {
-
-		File destinationDir = temporaryFolder.newFolder("maven");
-
-		return _buildTemplateWithMaven(
-			destinationDir, destinationDir, template, name, groupId, args);
-	}
-
-	private String _executeMaven(
-			File projectDir, boolean buildAndFail, String... args)
-		throws Exception {
-
-		File pomXmlFile = new File(projectDir, "pom.xml");
-
-		if (pomXmlFile.exists()) {
-			editXml(
-				pomXmlFile,
-				document -> {
-					addNexusRepositoriesElement(
-						document, "repositories", "repository");
-					addNexusRepositoriesElement(
-						document, "pluginRepositories", "pluginRepository");
-				});
-		}
-
-		String[] completeArgs = new String[args.length + 1];
-
-		completeArgs[0] = "--update-snapshots";
-
-		System.arraycopy(args, 0, completeArgs, 1, args.length);
-
-		MavenExecutor.Result result = mavenExecutor.execute(projectDir, args);
-
-		if (buildAndFail) {
-			Assert.assertFalse(
-				"Expected build to fail. " + result.exitCode,
-				result.exitCode == 0);
-		}
-		else {
-			Assert.assertEquals(result.output, 0, result.exitCode);
-		}
-
-		return result.output;
-	}
-
-	private String _executeMaven(File projectDir, String... args)
-		throws Exception {
-
-		return _executeMaven(projectDir, false, args);
 	}
 
 	private static URI _gradleDistribution;
