+++ modules/sdk/project-templates/project-templates-theme-contributor/src/main/resources/META-INF/services/com.liferay.project.templates.extensions.ProjectTemplateCustomizer
