+++ modules/sdk/project-templates/project-templates/src/test/java/com/liferay/project/templates/util/FileTestUtil.java
@@ -125,7 +125,9 @@ public class FileTestUtil {
 
 					String fileName = fileNamePath.toString();
 
-					if (fileName.startsWith(PROJECT_TEMPLATE_DIR_PREFIX)) {
+					if (fileName.startsWith(PROJECT_TEMPLATE_DIR_PREFIX) &&
+						!fileName.endsWith("-extensions")) {
+
 						return true;
 					}
